chore: min go version 1.18
